метрики работают

+++ src/NerualNetworkMLP/Model/GNN/GNN.cpp
@@ -52,9 +52,9 @@ void GraphNerualNetwork::train(Dataset& data,Dataset&  dataTest, double percentT
     auto elapsed_ms = std::chrono::duration_cast<std::chrono::seconds>(end - begin);
     _metrics.seconds=elapsed_ms.count();
     _metrics.accuracy =(_metrics.solutions.tp+_metrics.solutions.tn)/(data.getSize()*epoch);
-    _metrics.precision=_metrics.solutions.tp/(_metrics.solutions.tp+_metrics.solutions.fp);
-    _metrics.recall=_metrics.solutions.tp/(_metrics.solutions.tp+_metrics.solutions.fn);
-    _metrics.fMeasure=2*(_metrics.precision*_metrics.recall)/(_metrics.precision*_metrics.recall);
+    _metrics.precision=static_cast<double>(_metrics.solutions.tp)/(_metrics.solutions.tp+_metrics.solutions.fp);
+    _metrics.recall=static_cast<double>(_metrics.solutions.tp)/(_metrics.solutions.tp+_metrics.solutions.fn);
+    _metrics.fMeasure=2.0*(_metrics.precision*_metrics.recall)/(static_cast<double>(_metrics.precision)*_metrics.recall);
 }
 
 double GraphNerualNetwork::test(Dataset& data, double percentTestData) {
